+++ stdnet/apps/__init__.py
@@ -1,0 +1,4 @@
+'''\
+Collection of applications which may be relevant or not to the user.
+They show off the main features of the core library.
+'''