+++ stdnet/apps/grid/tests/models.py
@@ -1,0 +1,10 @@
+from stdnet import orm
+from stdnet.apps.grid import Queue
+
+
+class Task(orm.StdModel):
+    name = orm.CharField()
+    
+    
+class TaskQueue(Queue):
+    queue = orm.ListField(Task)