+++ stdnet/apps/timeseries/tests/regression/__init__.py
@@ -1,0 +1,2 @@
+from .hashtimeseries import *
+from .ts import *