+++ stdnet/contrib/timeseries/tests/regression/__init__.py
@@ -1,2 +1,0 @@
-from .hashtimeseries import *
-from .ts import *